backend/requirements.txt
aiofiles==25.1.0
anyio==4.11.0
async-property==0.2.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.3.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
python-keycloak==5.8.1
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
Flask-Cors==4.0.0